Cleanup refactor.
