+++ datacube/index/__init__.py
@@ -7,11 +7,14 @@ Modules for interfacing with the index/database.
 """
 
 from ._api import index_connect
+from ._spatial import strip_all_spatial_fields_from_query, extract_geom_from_query
 from .fields import UnknownFieldError
 from .exceptions import DuplicateRecordError, MissingRecordError, IndexSetupError
 from datacube.index.abstract import AbstractIndex as Index
 
 __all__ = [
+    'strip_all_spatial_fields_from_query',
+    'extract_geom_from_query',
     'index_connect',
     'Index',
