+++ datacube/model/_base.py
@@ -2,6 +2,7 @@
 #
 # Copyright (c) 2015-2024 ODC Contributors
 # SPDX-License-Identifier: Apache-2.0
+import datetime
 from collections import namedtuple
 
 Range = namedtuple('Range', ('begin', 'end'))
@@ -21,3 +22,5 @@ def ranges_overlap(ra: Range, rb: Range) -> bool:
 
 
 Not = namedtuple('Not', 'value')
+QueryField = str | float | int | Range | datetime.datetime | Not
+QueryDict = dict[str, QueryField]
